requirements.txt
fastapi
httpx
pydantic
python-dotenv
python-json-logger
pyyaml
uvicorn
